move exerciselist into model.exercise
