+++ src/main/java/seedu/zerotoone/MainApp.java
@@ -15,12 +15,12 @@ import seedu.zerotoone.commons.util.ConfigUtil;
 import seedu.zerotoone.commons.util.StringUtil;
 import seedu.zerotoone.logic.Logic;
 import seedu.zerotoone.logic.LogicManager;
-import seedu.zerotoone.model.ExerciseList;
 import seedu.zerotoone.model.Model;
 import seedu.zerotoone.model.ModelManager;
-import seedu.zerotoone.model.ReadOnlyExerciseList;
 import seedu.zerotoone.model.ReadOnlyUserPrefs;
 import seedu.zerotoone.model.UserPrefs;
+import seedu.zerotoone.model.exercise.ExerciseList;
+import seedu.zerotoone.model.exercise.ReadOnlyExerciseList;
 import seedu.zerotoone.model.util.SampleDataUtil;
 import seedu.zerotoone.storage.ExerciseListStorage;
 import seedu.zerotoone.storage.JsonExerciseListStorage;
