+++ src/main/java/seedu/zerotoone/logic/LogicManager.java
@@ -13,8 +13,8 @@ import seedu.zerotoone.logic.commands.exceptions.CommandException;
 import seedu.zerotoone.logic.parser.ParserManager;
 import seedu.zerotoone.logic.parser.exceptions.ParseException;
 import seedu.zerotoone.model.Model;
-import seedu.zerotoone.model.ReadOnlyExerciseList;
 import seedu.zerotoone.model.exercise.Exercise;
+import seedu.zerotoone.model.exercise.ReadOnlyExerciseList;
 import seedu.zerotoone.storage.Storage;
 
 /**
