+++ src/main/java/seedu/zerotoone/storage/Storage.java
@@ -5,9 +5,9 @@ import java.nio.file.Path;
 import java.util.Optional;
 
 import seedu.zerotoone.commons.exceptions.DataConversionException;
-import seedu.zerotoone.model.ReadOnlyExerciseList;
 import seedu.zerotoone.model.ReadOnlyUserPrefs;
 import seedu.zerotoone.model.UserPrefs;
+import seedu.zerotoone.model.exercise.ReadOnlyExerciseList;
 
 /**
  * API of the Storage component
