+++ src/main/java/seedu/zerotoone/storage/StorageManager.java
@@ -7,9 +7,9 @@ import java.util.logging.Logger;
 
 import seedu.zerotoone.commons.core.LogsCenter;
 import seedu.zerotoone.commons.exceptions.DataConversionException;
-import seedu.zerotoone.model.ReadOnlyExerciseList;
 import seedu.zerotoone.model.ReadOnlyUserPrefs;
 import seedu.zerotoone.model.UserPrefs;
+import seedu.zerotoone.model.exercise.ReadOnlyExerciseList;
 
 /**
  * Manages storage of ExerciseList data in local storage.
