+++ src/test/java/seedu/zerotoone/logic/LogicManagerTest.java
@@ -24,9 +24,9 @@ import seedu.zerotoone.logic.commands.exceptions.CommandException;
 import seedu.zerotoone.logic.parser.exceptions.ParseException;
 import seedu.zerotoone.model.Model;
 import seedu.zerotoone.model.ModelManager;
-import seedu.zerotoone.model.ReadOnlyExerciseList;
 import seedu.zerotoone.model.UserPrefs;
 import seedu.zerotoone.model.exercise.Exercise;
+import seedu.zerotoone.model.exercise.ReadOnlyExerciseList;
 import seedu.zerotoone.storage.JsonExerciseListStorage;
 import seedu.zerotoone.storage.JsonUserPrefsStorage;
 import seedu.zerotoone.storage.StorageManager;
