+++ src/test/java/seedu/zerotoone/testutil/TypicalExercises.java
@@ -4,8 +4,8 @@ import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.List;
 
-import seedu.zerotoone.model.ExerciseList;
 import seedu.zerotoone.model.exercise.Exercise;
+import seedu.zerotoone.model.exercise.ExerciseList;
 
 /**
  * A utility class containing a list of {@code Exercise} objects to be used in tests.
